+++ esa-20240910/alibabacloud_esa20240910/client.py
@@ -4273,7 +4273,7 @@ class Client(OpenApiClient):
         """
         @summary Add a New Load Balancer
         
-        @description Through this API, users can configure load balancing services according to their business needs, including but not limited to adaptive routing, weighted round-robin, rule matching, health checks, and other settings, to achieve effective traffic management and optimization.
+        @description Through this API, users can configure load balancing services according to their business needs, including but not limited to adaptive routing, weighted round-robin, rule matching, health checks, and more, to achieve effective traffic management and optimization.
         
         @param tmp_req: CreateLoadBalancerRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -4350,7 +4350,7 @@ class Client(OpenApiClient):
         """
         @summary Add a New Load Balancer
         
-        @description Through this API, users can configure load balancing services according to their business needs, including but not limited to adaptive routing, weighted round-robin, rule matching, health checks, and other settings, to achieve effective traffic management and optimization.
+        @description Through this API, users can configure load balancing services according to their business needs, including but not limited to adaptive routing, weighted round-robin, rule matching, health checks, and more, to achieve effective traffic management and optimization.
         
         @param tmp_req: CreateLoadBalancerRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -4426,7 +4426,7 @@ class Client(OpenApiClient):
         """
         @summary Add a New Load Balancer
         
-        @description Through this API, users can configure load balancing services according to their business needs, including but not limited to adaptive routing, weighted round-robin, rule matching, health checks, and other settings, to achieve effective traffic management and optimization.
+        @description Through this API, users can configure load balancing services according to their business needs, including but not limited to adaptive routing, weighted round-robin, rule matching, health checks, and more, to achieve effective traffic management and optimization.
         
         @param request: CreateLoadBalancerRequest
         @return: CreateLoadBalancerResponse
@@ -4441,7 +4441,7 @@ class Client(OpenApiClient):
         """
         @summary Add a New Load Balancer
         
-        @description Through this API, users can configure load balancing services according to their business needs, including but not limited to adaptive routing, weighted round-robin, rule matching, health checks, and other settings, to achieve effective traffic management and optimization.
+        @description Through this API, users can configure load balancing services according to their business needs, including but not limited to adaptive routing, weighted round-robin, rule matching, health checks, and more, to achieve effective traffic management and optimization.
         
         @param request: CreateLoadBalancerRequest
         @return: CreateLoadBalancerResponse